refactor: move builder class to entity
