--- src/main/java/kitchenpos/builder/MenuBuilder.java
@@ -1,56 +1,0 @@
-package kitchenpos.builder;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
-
-public class MenuBuilder {
-    private Long id;
-    private String name;
-    private BigDecimal price;
-    private Long menuGroupId;
-    private List<MenuProduct> menuProducts;
-
-    public MenuBuilder() {
-    }
-
-    public MenuBuilder(Long id, String name, BigDecimal price, Long menuGroupId,
-        List<MenuProduct> menuProducts) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.menuGroupId = menuGroupId;
-        this.menuProducts = menuProducts;
-    }
-
-    public MenuBuilder id(Long id) {
-        this.id = id;
-        return this;
-    }
-
-    public MenuBuilder name(String name) {
-        this.name = name;
-        return this;
-    }
-
-    public MenuBuilder price(BigDecimal price) {
-        this.price = price;
-        return this;
-    }
-
-    public MenuBuilder menuGroupId(Long menuGroupId) {
-        this.menuGroupId = menuGroupId;
-        return this;
-    }
-
-    public MenuBuilder menuProducts(List<MenuProduct> menuProducts) {
-        this.menuProducts = menuProducts;
-        return this;
-    }
-
-    public Menu build() {
-        return new Menu(id, name, price, menuGroupId, menuProducts);
-    }
-}
