+++ src/main/java/kitchenpos/domain/MenuGroup.java
@@ -6,13 +6,13 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import kitchenpos.builder.MenuGroupBuilder;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 @Entity
 public class MenuGroup {
@@ -26,4 +26,24 @@ public class MenuGroup {
     public static MenuGroupBuilder builder() {
         return new MenuGroupBuilder();
     }
+
+    public static class MenuGroupBuilder {
+        private Long id;
+        private String name;
+
+        public MenuGroupBuilder id(Long id) {
+            this.id = id;
+            return this;
+        }
+
+        public MenuGroupBuilder name(String name) {
+            this.name = name;
+            return this;
+        }
+
+        public MenuGroup build() {
+            return new MenuGroup(id, name);
+        }
+    }
+
 }
