+++ src/main/java/kitchenpos/domain/Order.java
@@ -10,13 +10,13 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
-import kitchenpos.builder.OrderBuilder;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 @Entity(name = "orders")
 public class Order {
@@ -42,5 +42,54 @@ public class Order {
 
     public OrderBuilder toBuilder() {
         return new OrderBuilder(id, orderTableId, orderStatus, orderedTime, orderLineItems);
+    }
+
+    public static class OrderBuilder {
+        private Long id;
+        private Long orderTableId;
+        private String orderStatus;
+        private LocalDateTime orderedTime;
+        private List<OrderLineItem> orderLineItems;
+
+        public OrderBuilder() {
+        }
+
+        public OrderBuilder(Long id, Long orderTableId, String orderStatus, LocalDateTime orderedTime,
+            List<OrderLineItem> orderLineItems) {
+            this.id = id;
+            this.orderTableId = orderTableId;
+            this.orderStatus = orderStatus;
+            this.orderedTime = orderedTime;
+            this.orderLineItems = orderLineItems;
+        }
+
+        public OrderBuilder id(Long id) {
+            this.id = id;
+            return this;
+        }
+
+        public OrderBuilder orderTableId(Long orderTableId) {
+            this.orderTableId = orderTableId;
+            return this;
+        }
+
+        public OrderBuilder orderStatus(String orderStatus) {
+            this.orderStatus = orderStatus;
+            return this;
+        }
+
+        public OrderBuilder orderedTime(LocalDateTime orderedTime) {
+            this.orderedTime = orderedTime;
+            return this;
+        }
+
+        public OrderBuilder orderLineItems(List<OrderLineItem> orderLineItems) {
+            this.orderLineItems = orderLineItems;
+            return this;
+        }
+
+        public Order build() {
+            return new Order(id, orderTableId, orderStatus, orderedTime, orderLineItems);
+        }
     }
 }
