+++ src/main/java/kitchenpos/domain/OrderTable.java
@@ -6,13 +6,13 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import kitchenpos.builder.OrderTableBuilder;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 @Entity
 public class OrderTable {
@@ -35,5 +35,46 @@ public class OrderTable {
 
     public OrderTableBuilder toBuilder() {
         return new OrderTableBuilder(id, tableGroupId, numberOfGuests, empty);
+    }
+
+    public static class OrderTableBuilder {
+        private Long id;
+        private Long tableGroupId;
+        private int numberOfGuests;
+        private boolean empty;
+
+        public OrderTableBuilder() {
+        }
+
+        public OrderTableBuilder(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
+            this.id = id;
+            this.tableGroupId = tableGroupId;
+            this.numberOfGuests = numberOfGuests;
+            this.empty = empty;
+        }
+
+        public OrderTableBuilder id(Long id) {
+            this.id = id;
+            return this;
+        }
+
+        public OrderTableBuilder tableGroupId(Long tableGroupId) {
+            this.tableGroupId = tableGroupId;
+            return this;
+        }
+
+        public OrderTableBuilder numberOfGuests(int numberOfGuests) {
+            this.numberOfGuests = numberOfGuests;
+            return this;
+        }
+
+        public OrderTableBuilder empty(boolean empty) {
+            this.empty = empty;
+            return this;
+        }
+
+        public OrderTable build() {
+            return new OrderTable(id, tableGroupId, numberOfGuests, empty);
+        }
     }
 }
